+++ templates/workspace_services/databricks/terraform/main.tf
@@ -7,6 +7,8 @@ resource "azurerm_databricks_workspace" "databricks" {
   infrastructure_encryption_enabled     = true
   public_network_access_enabled         = var.is_exposed_externally
   network_security_group_rules_required = "NoAzureDatabricksRules"
+  default_storage_firewall_enabled      = true
+  access_connector_id                   = azurerm_databricks_access_connector.connector.id
   tags                                  = local.tre_workspace_service_tags
 
   lifecycle { ignore_changes = [tags] }
@@ -25,4 +27,20 @@ resource "azurerm_databricks_workspace" "databricks" {
     azurerm_subnet_network_security_group_association.host,
     azurerm_subnet_network_security_group_association.container
   ]
+}
+
+// required to set default_storage_firewall_enabled = true
+// https://learn.microsoft.com/en-us/azure/databricks/security/network/storage/firewall-support
+//
+resource "azurerm_databricks_access_connector" "connector" {
+  name                = "${local.databricks_workspace_name}-access-connector"
+  resource_group_name = data.azurerm_resource_group.ws.name
+  location            = data.azurerm_resource_group.ws.location
+  tags                = local.tre_workspace_service_tags
+
+  identity {
+    type = "SystemAssigned"
+  }
+
+  lifecycle { ignore_changes = [tags] }
 }